requirements.txt
jinja2>=3.1
playwright>=1.40
pandas>=2.0
openpyxl>=3.1
pyyaml>=6.0
python-dotenv>=1.0
